Improved bidding logic

+++ Auktionssajt.Core/Services/BidService.cs
@@ -1,5 +1,6 @@
 using Auktionssajt.Core.Interfaces;
 using Auktionssajt.Data.Repository;
+using Auktionssajt.Domain.DTOs;
 using Auktionssajt.Domain.Entities;
 using Auktionssajt.Domain.Models;
 
@@ -8,26 +9,55 @@ namespace Auktionssajt.Core.Services
     public class BidService : IBidService
     {
         private readonly BidRepo _bidRepo = new();
+        private readonly AuctionRepo _auctionRepo = new();
         private readonly MappingService _mappingService = new();
+        private readonly UserService _userService = new();
 
-        public Status DeleteBid(int id)
+        public Status PlaceBid(NewBidModel newBid, int userId)
         {
+            var auction = _auctionRepo.GetAuction(newBid.AuctionID);
+            var bids = _bidRepo.GetBidsFromAuction(newBid.AuctionID);
+            if (auction.UserId == userId)
+                return Status.Forbidden;
+            
+            if (auction.EndTime < DateTime.Now)
+                return Status.Closed;
+
+            if (bids.Max(b => b.BidPrice) > newBid.BidPrice)
+                return Status.BidToLow;
+            
+            var bidEntity = _mappingService.ToBidEntity(newBid);
+            _bidRepo.InsertBid(bidEntity);
+
+            return Status.Ok;
+        }
+
+        public Status DeleteBid(int id, int userId)
+        {
+            var user = _userService.GetUser(userId);
+            var bid = _bidRepo.GetBid(id);
+
+            if (bid == null || user == null)
+                return Status.NotFound;
+
+            var auction = _auctionRepo.GetAuction(bid.AuctionId);
+            if (auction.EndTime > DateTime.Now)
+                return Status.Closed;
+            
+            if (bid.UserId != userId)
+                return Status.Unauthorized;
+            
             _bidRepo.DeleteBid(id);
             return Status.Ok;
         }
 
-        public List<BidEntity> GetBidList(int id)
+        public List<BidDTO> GetBids(int id)
         {
             var bids = _bidRepo.GetBidsFromAuction(id);
-            return bids;
-        }
-
-        public Status InsertBid(NewBidModel newBid)
-        {
-            var bidEntity = _mappingService.NewBidModeltoBidEntity(newBid);
-            _bidRepo.InsertBid(bidEntity);
-
-            return Status.Ok;
+            if (bids.Count < 1)
+                return null!;
+                
+            return _mappingService.ToBidDTO(bids);
         }
     }
 }
